requirements.txt
Pillow>=9.3.0
pytesseract>=0.3.10
uvicorn>=0.19.0
gunicorn
fastapi>=0.87.0
wheel
python-multipart
jinja2
numpy>=1.24
opencv-python>=4.7
python-dotenv
numpy
requests